Fixed missing name & fixed naming of urls. Removed User related resources.
