+++ impl/src/main/java/impl/AnswerResourceImpl.java
@@ -11,10 +11,14 @@ import com.github.seratch.jslack.api.model.block.composition.MarkdownTextObject;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import dao.AnswerDao;
+import dao.AnswerVote;
+import dao.AnswerVoteDao;
 import dao.QuestionDao;
+import io.netty.handler.codec.http.HttpResponseStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import rx.Observable;
+import rx.functions.Func1;
 import se.fortnox.reactivewizard.db.transactions.DaoTransactions;
 import se.fortnox.reactivewizard.jaxrs.WebException;
 import slack.SlackResource;
@@ -37,6 +41,13 @@ public class AnswerResourceImpl implements AnswerResource {
 
     public static final String ERROR_NOT_OWNER_OF_QUESTION = "not.owner.of.question";
     public static final String ERROR_ANSWER_NOT_CREATED    = "answer.not.created";
+    public static final String ANSWER_NOT_FOUND                    = "answer.not.found";
+    public static final String NOT_OWNER_OF_ANSWER                 = "not.owner.of.answer";
+    public static final String FAILED_TO_UPDATE_ANSWER             = "failed.to.update.answer";
+    public static final String FAILED_TO_DELETE_ANSWER             = "failed.to.delete.answer";
+    public static final String FAILED_TO_GET_ANSWER_FROM_DATABASE  = "failed.to.get.answer.from.database";
+    public static final String FAILED_TO_GET_ANSWERS_FROM_DATABASE = "failed.to.get.answers.from.database";
+    public static final String INVALID_VOTE = "invalid.vote";
 
     private static final Logger LOG = LoggerFactory.getLogger(AnswerResourceImpl.class);
 
@@ -46,6 +57,7 @@ public class AnswerResourceImpl implements AnswerResource {
     private final SlackResource     slackResource;
     private final UserResource      userResource;
     private final ApplicationConfig applicationConfig;
+    private final AnswerVoteDao     answerVoteDao;
 
     @Inject
     public AnswerResourceImpl(AnswerDao answerDao,
@@ -53,7 +65,9 @@ public class AnswerResourceImpl implements AnswerResource {
         DaoTransactions daoTransactions,
         SlackResource slackResource,
         UserResource userResource,
-        ApplicationConfig applicationConfig
+        ApplicationConfig applicationConfig,
+        AnswerVoteDao answerVoteDao
+
     ) {
         this.answerDao = answerDao;
         this.questionDao = questionDao;
@@ -61,10 +75,12 @@ public class AnswerResourceImpl implements AnswerResource {
         this.slackResource = slackResource;
         this.userResource = userResource;
         this.applicationConfig = applicationConfig;
+        this.answerVoteDao = answerVoteDao;
+
     }
 
     @Override
-    public Observable<Answer> answerQuestion(Auth auth, Answer answer, long questionId) {
+    public Observable<Answer> createAnswer(Auth auth, Answer answer, long questionId) {
         Objects.requireNonNull(answer.getAnswer());
 
         return this.answerDao.createAnswer(auth.getUserId(), questionId, answer)
@@ -140,7 +156,8 @@ public class AnswerResourceImpl implements AnswerResource {
             .toList()
             .doOnError(throwable -> {
                 LOG.error("Failed to get answers for question: " + questionId, throwable);
-            });
+            })
+            .onErrorResumeNext((e) -> error(new WebException(INTERNAL_SERVER_ERROR, FAILED_TO_GET_ANSWERS_FROM_DATABASE, e)));
     }
 
     @Override
@@ -156,6 +173,45 @@ public class AnswerResourceImpl implements AnswerResource {
     }
 
     @Override
+    public Observable<Void> updateAnswer(Auth auth, long answerId, Answer answer) {
+        return answerDao.getAnswerById(answerId)
+            .onErrorResumeNext(throwable -> error(new WebException(INTERNAL_SERVER_ERROR, FAILED_TO_GET_ANSWER_FROM_DATABASE, throwable)))
+            .switchIfEmpty(exception(() -> new WebException(HttpResponseStatus.NOT_FOUND, ANSWER_NOT_FOUND)))
+            .flatMap(storedAnswer -> {
+                if (auth.getUserId() != storedAnswer.getUserId()) {
+                    return error(new WebException(HttpResponseStatus.FORBIDDEN, NOT_OWNER_OF_ANSWER));
+                }
+                return answerDao.updateAnswer(auth.getUserId(), answerId, answer)
+                    .onErrorResumeNext(throwable -> error(new WebException(INTERNAL_SERVER_ERROR, FAILED_TO_UPDATE_ANSWER, throwable)));
+            });
+    }
+
+    @Override
+    public Observable<Void> deleteAnswer(Auth auth, long answerId) {
+        return answerDao.getAnswerById(answerId)
+            .onErrorResumeNext(throwable -> error(new WebException(INTERNAL_SERVER_ERROR, FAILED_TO_GET_ANSWER_FROM_DATABASE, throwable)))
+            .switchIfEmpty(exception(() -> new WebException(HttpResponseStatus.NOT_FOUND, ANSWER_NOT_FOUND)))
+            .flatMap(storedAnswer -> {
+                if (auth.getUserId() != storedAnswer.getUserId()) {
+                    return error(new WebException(HttpResponseStatus.FORBIDDEN, NOT_OWNER_OF_ANSWER));
+                }
+                return answerDao.deleteAnswer(auth.getUserId(), answerId)
+                    .onErrorResumeNext(throwable -> error(new WebException(INTERNAL_SERVER_ERROR, FAILED_TO_DELETE_ANSWER, throwable)));
+            });
+    }
+
+    @Override
+    public Observable<Void> upVoteAnswer(Auth auth, long answerId) {
+        return handleVote(new AnswerVote(auth.getUserId(), answerId, 1));
+    }
+
+    @Override
+    public Observable<Void> downVoteAnswer(Auth auth, long answerId) {
+        return handleVote(new AnswerVote(auth.getUserId(), answerId, -1));
+    }
+
+
+    @Override
     public Observable<Void> markAsAcceptedAnswer(Auth auth, long answerId) {
         return answerDao.getAnswerById(answerId)
             .flatMap(answer -> {
@@ -166,5 +222,35 @@ public class AnswerResourceImpl implements AnswerResource {
                 Observable<Integer> markQuestionAsAnswered = questionDao.markAsAnswered(auth.getUserId(), answer.getQuestionId());
                 return daoTransactions.executeTransaction(markAnswerAsAccepted, markQuestionAsAnswered);
             });
+    }
+
+
+    private Observable<Void> handleVote(AnswerVote newVote) {
+        return answerDao.getAnswerById(newVote.getAnswerId())
+            .flatMap(validateAnswerAndGetExistingVote(newVote))
+            .flatMap(validateVoteAndRemoveIfZero(newVote))
+            .switchIfEmpty(answerVoteDao.createVote(newVote).map(i -> newVote))
+            .ignoreElements()
+            .cast(Void.class);
+    }
+
+    private Func1<Answer, Observable<AnswerVote>> validateAnswerAndGetExistingVote(AnswerVote newVote) {
+        return answer -> {
+            if (answer.getUserId() == newVote.getUserId()) { // no voting for your own answer
+                return error(new WebException(BAD_REQUEST, INVALID_VOTE));
+            }
+            return answerVoteDao.findVote(newVote.getUserId(), newVote.getAnswerId());
+        };
+    }
+
+    private Func1<AnswerVote, Observable<AnswerVote>> validateVoteAndRemoveIfZero(AnswerVote newVote) {
+        return existingVote -> {
+            int totalVote = newVote.getValue() + existingVote.getValue();
+            if (totalVote == 0) {
+                return answerVoteDao.deleteVote(existingVote.getUserId(), existingVote.getAnswerId())
+                    .map(i -> existingVote);
+            }
+            return error(new WebException(BAD_REQUEST, INVALID_VOTE));
+        };
     }
 }
