+++ impl/src/test/java/impl/AnswerResourceImplTest.java
@@ -1,40 +1,45 @@
 package impl;
 
-import api.UserAnswerResource;
+import api.AnswerResource;
+import api.UserResource;
 import api.auth.Auth;
 import dao.AnswerDao;
 import dao.AnswerInternal;
 import dao.AnswerVoteDao;
+import dao.QuestionDao;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
+import se.fortnox.reactivewizard.db.transactions.DaoTransactions;
 import se.fortnox.reactivewizard.jaxrs.WebException;
+import slack.SlackResource;
 
 import java.sql.SQLException;
 
-import static impl.UserAnswerResourceImpl.ANSWER_NOT_FOUND;
-import static impl.UserAnswerResourceImpl.FAILED_TO_DELETE_ANSWER;
-import static impl.UserAnswerResourceImpl.FAILED_TO_GET_ANSWERS_FROM_DATABASE;
-import static impl.UserAnswerResourceImpl.FAILED_TO_GET_ANSWER_FROM_DATABASE;
-import static impl.UserAnswerResourceImpl.FAILED_TO_UPDATE_ANSWER;
-import static impl.UserAnswerResourceImpl.NOT_OWNER_OF_ANSWER;
+import static impl.AnswerResourceImpl.ANSWER_NOT_FOUND;
+import static impl.AnswerResourceImpl.FAILED_TO_DELETE_ANSWER;
+import static impl.AnswerResourceImpl.FAILED_TO_GET_ANSWERS_FROM_DATABASE;
+import static impl.AnswerResourceImpl.FAILED_TO_GET_ANSWER_FROM_DATABASE;
+import static impl.AnswerResourceImpl.FAILED_TO_UPDATE_ANSWER;
+import static impl.AnswerResourceImpl.NOT_OWNER_OF_ANSWER;
 import static io.netty.handler.codec.http.HttpResponseStatus.FORBIDDEN;
 import static io.netty.handler.codec.http.HttpResponseStatus.INTERNAL_SERVER_ERROR;
 import static io.netty.handler.codec.http.HttpResponseStatus.NOT_FOUND;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 import static rx.Observable.empty;
 import static rx.Observable.error;
 import static rx.Observable.just;
 
-public class UserAnswerResourceImplTest {
+public class AnswerResourceImplTest {
 
-    private UserAnswerResource userAnswerResource;
-    private Auth auth;
+    private AnswerResource answerResource;
+    private Auth           auth;
 
     @Mock
     private AnswerDao          answerDao;
@@ -45,7 +50,8 @@ public class UserAnswerResourceImplTest {
     @Before
     public void beforeEach() {
         initMocks(this);
-        userAnswerResource = new UserAnswerResourceImpl(answerDao, answerVoteDao);
+        ApplicationConfig applicationConfig = new ApplicationConfig();
+        answerResource = new AnswerResourceImpl(answerDao, mock(QuestionDao.class), mock(DaoTransactions.class), mock(SlackResource.class), mock(UserResource.class), applicationConfig, answerVoteDao);
         auth = new Auth();
         auth.setUserId(123);
 
@@ -53,9 +59,9 @@ public class UserAnswerResourceImplTest {
 
     @Test
     public void shouldThrowInternalServerErrorIfGetAnswersFails() {
-        when(answerDao.getAnswers(123, 123)).thenReturn(error(new SQLException("poff")));
+        when(answerDao.getAnswersWithUserVotes( auth.getUserId(), 123)).thenReturn(error(new SQLException("poff")));
 
-        assertException(() -> userAnswerResource.getAnswers(123, 123).toBlocking().singleOrDefault(null),
+        assertException(() -> answerResource.getAnswers(auth, 123).toBlocking().singleOrDefault(null),
             INTERNAL_SERVER_ERROR,
             FAILED_TO_GET_ANSWERS_FROM_DATABASE);
     }
@@ -66,7 +72,7 @@ public class UserAnswerResourceImplTest {
         when(answerDao.updateAnswer(123,  123, answer)).thenReturn(error(new SQLException("poff")));
         when(answerDao.getAnswerById(123)).thenReturn(just(answer));
 
-        assertException(() -> userAnswerResource.updateAnswer(auth, 123,answer).toBlocking().singleOrDefault(null),
+        assertException(() -> answerResource.updateAnswer(auth, 123,answer).toBlocking().singleOrDefault(null),
             INTERNAL_SERVER_ERROR,
             FAILED_TO_UPDATE_ANSWER);
     }
@@ -78,7 +84,7 @@ public class UserAnswerResourceImplTest {
         when(answerDao.updateAnswer(123,  123, answer)).thenReturn(error(new SQLException("poff")));
         when(answerDao.getAnswerById(123)).thenReturn(just(answer));
 
-        assertException(() -> userAnswerResource.updateAnswer(auth, 123,answer).toBlocking().singleOrDefault(null),
+        assertException(() -> answerResource.updateAnswer(auth, 123,answer).toBlocking().singleOrDefault(null),
             FORBIDDEN,
             NOT_OWNER_OF_ANSWER);
     }
@@ -87,7 +93,7 @@ public class UserAnswerResourceImplTest {
     public void shouldThrowNotFoundIfAnswerToUpdateCannotBeFound() {
         when(answerDao.getAnswerById(123)).thenReturn(empty());
 
-        assertException(() -> userAnswerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
+        assertException(() -> answerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
             NOT_FOUND,
             ANSWER_NOT_FOUND);
     }
@@ -97,7 +103,7 @@ public class UserAnswerResourceImplTest {
         AnswerInternal answer = createAnswer(444);
         when(answerDao.getAnswerById(123)).thenReturn(just(answer));
 
-        assertException(() -> userAnswerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
+        assertException(() -> answerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
             FORBIDDEN,
             NOT_OWNER_OF_ANSWER);
     }
@@ -106,7 +112,7 @@ public class UserAnswerResourceImplTest {
     public void shouldThrowNotFoundIfAnswerToDeleteCannotBeFound() {
         when(answerDao.getAnswerById(123)).thenReturn(empty());
 
-        assertException(() -> userAnswerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
+        assertException(() -> answerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
             NOT_FOUND,
             ANSWER_NOT_FOUND);
     }
@@ -117,7 +123,7 @@ public class UserAnswerResourceImplTest {
         when(answerDao.deleteAnswer(123,  123)).thenReturn(error(new SQLException("poff")));
         when(answerDao.getAnswerById(123)).thenReturn(just(answer));
 
-        assertException(() -> userAnswerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
+        assertException(() -> answerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
             INTERNAL_SERVER_ERROR,
             FAILED_TO_DELETE_ANSWER);
     }
@@ -128,7 +134,7 @@ public class UserAnswerResourceImplTest {
         Auth auth = new Auth();
         auth.setUserId(123);
 
-        assertException(() -> userAnswerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
+        assertException(() -> answerResource.deleteAnswer(auth, 123).toBlocking().singleOrDefault(null),
             INTERNAL_SERVER_ERROR,
             FAILED_TO_GET_ANSWER_FROM_DATABASE);
     }
