+++ impl/src/test/java/impl/QuestionResourceImplTest.java
@@ -4,6 +4,9 @@ import api.Question;
 import api.QuestionResource;
 import api.auth.Auth;
 import dao.QuestionDao;
+import dao.QuestionVoteDao;
+import io.netty.handler.codec.http.HttpResponseStatus;
+import org.assertj.core.api.ThrowableAssert;
 import org.junit.Before;
 import org.junit.Test;
 import rx.Observable;
@@ -11,22 +14,36 @@ import se.fortnox.reactivewizard.jaxrs.WebException;
 
 import java.sql.SQLException;
 
+import static impl.QuestionResourceImpl.FAILED_TO_DELETE_QUESTION;
+import static impl.QuestionResourceImpl.FAILED_TO_GET_QUESTIONS_FROM_DATABASE;
+import static impl.QuestionResourceImpl.FAILED_TO_GET_QUESTION_FROM_DATABASE;
+import static impl.QuestionResourceImpl.FAILED_TO_UPDATE_QUESTION_TO_DATABASE;
+import static impl.QuestionResourceImpl.NOT_OWNER_OF_QUESTION;
+import static io.netty.handler.codec.http.HttpResponseStatus.FORBIDDEN;
 import static io.netty.handler.codec.http.HttpResponseStatus.INTERNAL_SERVER_ERROR;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+import static rx.Observable.error;
+import static rx.Observable.just;
 
 public class QuestionResourceImplTest {
 
     private QuestionResource questionResource;
     private QuestionDao          questionDao;
-
+    private QuestionVoteDao      questionVoteDao;
+    private Question question;
+    private Auth auth;
     @Before
     public void beforeEach() {
         questionDao = mock(QuestionDao.class);
-        questionResource = new QuestionResourceImpl(questionDao);
+        questionVoteDao = mock(QuestionVoteDao.class);
+        questionResource = new QuestionResourceImpl(questionDao, questionVoteDao);
+        auth = new Auth(123);
+        question = createQuestion(123);
     }
+
 
     @Test
     public void shouldReturnInternalServerErrorWhenPostQuestionFails() {
@@ -46,5 +63,83 @@ public class QuestionResourceImplTest {
             });
     }
 
+
+    @Test
+    public void shouldReturnInternalServerErrorWhenGetQuestionsFails() {
+        when(questionDao.getQuestions(123)).thenReturn(Observable.error(new SQLException("poff")));
+
+        assertException(() -> questionResource.getQuestions( 123).toBlocking().singleOrDefault(null),
+            INTERNAL_SERVER_ERROR,
+            FAILED_TO_GET_QUESTIONS_FROM_DATABASE);
+    }
+
+    @Test
+    public void shouldReturnInternalServerErrorWhenUpdateQuestionFails() {
+
+        when(questionDao.getQuestion(123)).thenReturn(just(question));
+        when(questionDao.updateQuestion(123, 123, question)).thenReturn(Observable.error(new SQLException("poff")));
+
+        assertException(() -> questionResource.updateQuestion(auth, 123, question).toBlocking().singleOrDefault(null),
+            INTERNAL_SERVER_ERROR,
+            FAILED_TO_UPDATE_QUESTION_TO_DATABASE);
+    }
+
+    @Test
+    public void shouldThrowInternalServerErrorIfQuestionCannotBeFetchedOnUpdate() {
+        when(questionDao.getQuestion(123))
+            .thenReturn(just(question))
+            .thenReturn(error(new SQLException("poff")));
+        when(questionDao.updateQuestion(123, 123, question)).thenReturn(just(1));
+
+        assertException(() -> questionResource.updateQuestion(auth, 123, question).toBlocking().singleOrDefault(null),
+            INTERNAL_SERVER_ERROR,
+            FAILED_TO_GET_QUESTION_FROM_DATABASE);
+    }
+
+    @Test
+    public void shouldThrowInternalServerErrorIfQuestionCannotBeFetchedAfterUpdate() {
+        when(questionDao.getQuestion(123)).thenReturn(Observable.error(new SQLException("poff")));
+
+        assertException(() -> questionResource.updateQuestion(auth, 123, question).toBlocking().singleOrDefault(null),
+            INTERNAL_SERVER_ERROR,
+            FAILED_TO_GET_QUESTION_FROM_DATABASE);
+    }
+
+    @Test
+    public void shouldThrowForbiddenIfQuestionIsNotCreatedByTheDeleter() {
+        Question question = createQuestion(444);
+        when(questionDao.getQuestion(123)).thenReturn(just(question));
+
+        assertException(() -> questionResource.deleteQuestion(auth, 123).toBlocking().singleOrDefault(null),
+            FORBIDDEN,
+            NOT_OWNER_OF_QUESTION);
+    }
+
+
+    @Test
+    public void shouldThrowInternalIfQuestionToDeleteCannotBeDeleted() {
+
+        when(questionDao.deleteQuestion(123,  123)).thenReturn(Observable.error(new SQLException("poff")));
+        when(questionDao.getQuestion(123)).thenReturn(just(question));
+
+        assertException(() -> questionResource.deleteQuestion(auth, 123).toBlocking().singleOrDefault(null),
+            INTERNAL_SERVER_ERROR,
+            FAILED_TO_DELETE_QUESTION);
+    }
+
+    private static void assertException(ThrowableAssert.ThrowingCallable observable, HttpResponseStatus responseStatus, String error) {
+        assertThatExceptionOfType(WebException.class)
+            .isThrownBy(observable)
+            .satisfies(e -> {
+                assertEquals(responseStatus, e.getStatus());
+                assertEquals(error, e.getError());
+            });
+    }
+
+    private Question createQuestion(long userId) {
+        Question question = new Question();
+        question.setUserId(userId);
+        return question;
+    }
 
 }
