+++ spec/src/test/java/impl/AnswerResourceTest.java
@@ -20,15 +20,19 @@ import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.testcontainers.containers.PostgreSQLContainer;
+import rx.Observable;
 import se.fortnox.reactivewizard.db.transactions.DaoTransactions;
 import se.fortnox.reactivewizard.jaxrs.WebException;
 import slack.SlackResource;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.BiFunction;
 
+import static impl.AnswerResourceImpl.ANSWER_NOT_FOUND;
 import static impl.AnswerResourceImpl.ERROR_ANSWER_NOT_CREATED;
 import static impl.AnswerResourceImpl.ERROR_NOT_OWNER_OF_QUESTION;
+import static impl.AnswerResourceImpl.INVALID_VOTE;
 import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
 import static io.netty.handler.codec.http.HttpResponseStatus.INTERNAL_SERVER_ERROR;
 import static io.netty.handler.codec.http.HttpResponseStatus.NOT_FOUND;
@@ -39,6 +43,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.tuple;
 import static org.assertj.core.api.Assertions.within;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyListOf;
 import static org.mockito.Matchers.anyLong;
@@ -80,6 +86,7 @@ public class AnswerResourceTest {
         userResource = testSetup.getInjector().getInstance(UserResource.class);
         mockedSlackResource = testSetup.getInjector().getInstance(SlackResource.class);
         answerVoteDao = testSetup.getInjector().getInstance(AnswerVoteDao.class);
+
     }
 
     @After
@@ -110,7 +117,7 @@ public class AnswerResourceTest {
         Auth answerer = newUser();
         // and someone answers the question
         Answer answer         = newAnswer();
-        Answer returnedAnswer = answerResource.answerQuestion(answerer, answer, returnedQuestion.getId()).toBlocking().singleOrDefault(null);
+        Answer returnedAnswer = answerResource.createAnswer(answerer, answer, returnedQuestion.getId()).toBlocking().singleOrDefault(null);
         assertThat(returnedAnswer).isNotNull();
 
         //Verify slack notification is send to the user who created the question
@@ -164,7 +171,7 @@ public class AnswerResourceTest {
 
         // and someone answers the question
         Answer answer         = newAnswer();
-        Answer returnedAnswer = answerResource.answerQuestion(questioner, answer, returnedQuestion.getId()).toBlocking().singleOrDefault(null);
+        Answer returnedAnswer = answerResource.createAnswer(questioner, answer, returnedQuestion.getId()).toBlocking().singleOrDefault(null);
         assertThat(returnedAnswer).isNotNull();
 
         //Verify slack notification is send to the user who created the question
@@ -180,7 +187,7 @@ public class AnswerResourceTest {
         // and an answer
         Answer answer         = newAnswer();
         // when the current user tries to accept the question, that belongs to someone else
-        Answer returnedAnswer = answerResource.answerQuestion(newUser(), answer, returnedQuestion.getId()).toBlocking().singleOrDefault(null);
+        Answer returnedAnswer = answerResource.createAnswer(newUser(), answer, returnedQuestion.getId()).toBlocking().singleOrDefault(null);
         // then a exception shall be thrown stating that the request is invalid.
         assertThatExceptionOfType(WebException.class)
             .isThrownBy(() -> answerResource.markAsAcceptedAnswer(newUser(), returnedAnswer.getId()).toBlocking().singleOrDefault(null))
@@ -198,11 +205,18 @@ public class AnswerResourceTest {
         RuntimeException dbException   = new RuntimeException();
         AnswerDao        answerDaoMock = mock(AnswerDao.class);
         when(answerDaoMock.createAnswer(anyLong(), anyLong(), any())).thenReturn(error(dbException));
-        AnswerResource mockedResource = new AnswerResourceImpl(answerDaoMock, null, testSetup.getInjector().getInstance(DaoTransactions.class), mockedSlackResource, userResource, new ApplicationConfig());
+        AnswerResource mockedResource = new AnswerResourceImpl(
+            answerDaoMock,
+            null,
+            testSetup.getInjector().getInstance(DaoTransactions.class),
+            mockedSlackResource,
+            userResource,
+            new ApplicationConfig(),
+            null);
 
         // when the answer is going to be persisted, exception is returned
         assertThatExceptionOfType(WebException.class)
-            .isThrownBy(() -> mockedResource.answerQuestion(newUser(), newAnswer(), question.getId()).toBlocking().singleOrDefault(null))
+            .isThrownBy(() -> mockedResource.createAnswer(newUser(), newAnswer(), question.getId()).toBlocking().singleOrDefault(null))
             .withCause(dbException)
             .satisfies(e -> {
                 assertThat(e.getStatus()).isEqualTo(INTERNAL_SERVER_ERROR);
@@ -214,7 +228,7 @@ public class AnswerResourceTest {
     public void shouldReturnAnswerByIdWithCorrectVotes() {
 
         Question question = questionResource.createQuestion(newUser(), newQuestion()).toBlocking().singleOrDefault(null);
-        Answer answer = answerResource.answerQuestion(newUser(), newAnswer(), question.getId()).toBlocking().singleOrDefault(null);
+        Answer answer = answerResource.createAnswer(newUser(), newAnswer(), question.getId()).toBlocking().singleOrDefault(null);
         Auth user = newUser();
 
         assertAnswerById(answer, 0);
@@ -248,7 +262,7 @@ public class AnswerResourceTest {
         Auth user3 = newUser();
 
         Question question = questionResource.createQuestion(user1, newQuestion()).toBlocking().singleOrDefault(null);
-        Answer answer1 = answerResource.answerQuestion(user1, newAnswer(), question.getId()).toBlocking().singleOrDefault(null);
+        Answer answer1 = answerResource.createAnswer(user1, newAnswer(), question.getId()).toBlocking().singleOrDefault(null);
 
         assertVotesByQuestionId(user1, question, tuple(0, 0));
 
@@ -262,12 +276,210 @@ public class AnswerResourceTest {
         addVote(user3, answer1, -1);
         assertVotesByQuestionId(user3, question, tuple(1, -1));
 
-        Answer answer2 = answerResource.answerQuestion(user1, newAnswer(), question.getId()).toBlocking().singleOrDefault(null);
+        Answer answer2 = answerResource.createAnswer(user1, newAnswer(), question.getId()).toBlocking().singleOrDefault(null);
         assertVotesByQuestionId(user3, question, tuple(1, -1), tuple(0, 0));
 
         addVote(user3, answer2, -1);
         assertVotesByQuestionId(user3, question, tuple(1, -1), tuple(-1, -1));
         assertVotesByQuestionId(user1, question, tuple(1, 1), tuple(-1, 0));
+    }
+
+    @Test
+    public void shouldAddAnswerToQuestion() {
+
+        Auth createdUserAuth = newUser();
+        Question question = createQuestion(createdUserAuth);
+
+        Answer answer = createAnswer();
+
+        answerResource.createAnswer(createdUserAuth, answer, question.getId()).toBlocking().singleOrDefault(null);
+
+        Answer createdAnswer = getFirstAnswer(createdUserAuth, question);
+        assertFalse(createdAnswer.isAccepted());
+        assertEquals("this is the body of the answer", createdAnswer.getAnswer());
+        assertThat(createdAnswer.getVotes()).isZero();
+        assertEquals("Test Subject", createdAnswer.getCreatedBy());
+    }
+
+    @Test
+    public void shouldBeAbleToUpdateAnswer() {
+
+        Auth createdUserAuth = newUser();
+        Question question = createQuestion(createdUserAuth);
+
+        Answer answer = new Answer();
+        answer.setAnswer("this is the body of the answer");
+
+        answerResource.createAnswer(createdUserAuth, answer, question.getId()).toBlocking().singleOrDefault(null);
+
+        Answer updatedAnswerInput = new Answer();
+        updatedAnswerInput.setAnswer("this is an updated body of the answer");
+
+        answerResource.updateAnswer(createdUserAuth, answer.getId(), updatedAnswerInput).toBlocking().singleOrDefault(null);
+
+        Answer updatedAnswer = getFirstAnswer(createdUserAuth, question);
+
+        assertFalse(updatedAnswer.isAccepted());
+        assertEquals("this is an updated body of the answer", updatedAnswer.getAnswer());
+        assertEquals("Test Subject", updatedAnswer.getCreatedBy());
+    }
+
+    @Test
+    public void shouldBeAbleToDeleteAnswer() {
+        // given answers exists
+        Auth auth = newUser();
+        Question question = createQuestion(auth);
+
+        Answer answer = createAnswer();
+        Answer answer2 = createAnswer("title2","answer2");
+
+        answerResource.createAnswer(auth, answer, question.getId()).toBlocking().singleOrDefault(null);
+        answerResource.createAnswer(auth, answer2, question.getId()).toBlocking().singleOrDefault(null);
+
+        Answer createdAnswer = getFirstAnswer(auth, question);
+
+        // when we delete a answer
+        answerResource.deleteAnswer(auth,  createdAnswer.getId()).toBlocking().singleOrDefault(null);
+
+        // then only the answer we want to delete should be deleted
+        List<Answer> remainingAnswers = answerResource.getAnswers(auth, question.getId()).toBlocking().single();
+        assertThat(remainingAnswers.size()).isEqualTo(1);
+    }
+
+    @Test
+    public void shouldNotDeleteAnswerThatDoesNotExist() {
+        Auth auth = newUser();
+        long nonExistingAnswerId = 12;
+        assertThatExceptionOfType(WebException.class)
+            .isThrownBy(() -> answerResource.deleteAnswer(auth,  nonExistingAnswerId).toBlocking().singleOrDefault(null))
+            .satisfies(e -> {
+                assertEquals(NOT_FOUND, e.getStatus());
+                assertEquals(ANSWER_NOT_FOUND, e.getError());
+            });
+    }
+
+    @Test
+    public void shouldBeAbleToFetchAnswersForGivenUserAndQuestion() {
+
+        Auth authUserOne = newUser();
+        Question question = createQuestion(authUserOne);
+
+        Answer answer = createAnswer();
+
+        Auth authUserTwo = newUser();
+
+        Question questionForUserTwo = createQuestion(authUserTwo);
+
+        Answer answerForUserTwo = createAnswer();
+        answerResource.createAnswer(authUserOne, answer, question.getId()).toBlocking().singleOrDefault(null);
+        answerResource.createAnswer(authUserTwo, answerForUserTwo, questionForUserTwo.getId()).toBlocking().singleOrDefault(null);
+
+        Answer createdAnswer = getFirstAnswer(authUserOne, question);
+        assertFalse(createdAnswer.isAccepted());
+        assertEquals("this is the body of the answer", createdAnswer.getAnswer());
+        assertThat(createdAnswer.getVotes()).isZero();
+        assertEquals("Test Subject", createdAnswer.getCreatedBy());
+    }
+
+    @Test
+    public void shouldYieldCorrectVotesWhenDownVotingAndUpVoting() {
+
+        Auth auth1 = newUser();
+        Auth auth2 = newUser();
+        Auth auth3 = newUser();
+        Answer answer = createQuestionAndAnswer(newUser());
+
+        voteAndAssertSuccess(answerResource::downVoteAnswer, auth1, answer,-1);
+        voteAndAssertFailure(answerResource::downVoteAnswer, auth1, answer);
+        voteAndAssertSuccess(answerResource::upVoteAnswer, auth1, answer, 0);
+        assertNoVote(auth1, answer);
+        voteAndAssertSuccess(answerResource::upVoteAnswer, auth1, answer, 1);
+        voteAndAssertFailure(answerResource::upVoteAnswer, auth1, answer);
+
+        voteAndAssertSuccess(answerResource::upVoteAnswer, auth2, answer, 2);
+
+        voteAndAssertSuccess(answerResource::upVoteAnswer, auth3, answer, 3);
+
+        voteAndAssertSuccess(answerResource::downVoteAnswer, auth2, answer, 2);
+        assertNoVote(auth2, answer);
+
+        voteAndAssertSuccess(answerResource::downVoteAnswer, auth3, answer, 1);
+        assertNoVote(auth2, answer);
+    }
+
+    @Test
+    public void shouldThrowErrorWhenUserVotesOnOwnAnswer() {
+
+        Auth auth = newUser();
+        Answer answer = createQuestionAndAnswer(auth);
+
+        voteAndAssertFailure(answerResource::downVoteAnswer, auth, answer);
+        voteAndAssertFailure(answerResource::upVoteAnswer, auth, answer);
+    }
+
+    private void voteAndAssertSuccess(BiFunction<Auth, Long, Observable<Void>> call, Auth auth, Answer answer, int expectedVotes) {
+        call.apply(auth, answer.getId()).test()
+            .awaitTerminalEvent()
+            .assertNoErrors();
+        assertThat(answerResource.getAnswerById(answer.getId()).test().awaitTerminalEvent().getOnNextEvents())
+            .extracting(Answer::getVotes)
+            .containsExactly(expectedVotes);
+    }
+
+    private void voteAndAssertFailure(BiFunction<Auth, Long, Observable<Void>> call, Auth auth, Answer answer) {
+        assertThatExceptionOfType(WebException.class)
+            .isThrownBy(() -> call.apply(auth, answer.getId()).toBlocking().single())
+            .satisfies(e -> {
+                assertThat(e).isInstanceOf(WebException.class);
+                assertThat(e).hasFieldOrPropertyWithValue("status", BAD_REQUEST);
+                assertThat(e).hasFieldOrPropertyWithValue("error", INVALID_VOTE);
+            });
+    }
+
+    private void assertNoVote(Auth auth, Answer answer) {
+        answerVoteDao.findVote(auth.getUserId(), answer.getId()).test().awaitTerminalEvent()
+            .assertNoValues();
+    }
+
+    private Answer createQuestionAndAnswer(Auth auth) {
+        Question question = createQuestion(auth);
+        Answer answer = new Answer();
+        answer.setAnswer("answer");
+        return answerResource.createAnswer(auth, answer, question.getId()).toBlocking().singleOrDefault(null);
+    }
+
+    private Answer getFirstAnswer(Auth auth, Question question) {
+        List<Answer> createdAnswers = answerResource.getAnswers(auth, question.getId()).toBlocking().single();
+        assertThat(createdAnswers).isNotEmpty();
+        return createdAnswers.get(0);
+    }
+
+    private Question createQuestion(Auth auth) {
+        // when question is inserted
+        Question question = new Question();
+        question.setAnswerAccepted(false);
+        question.setBounty(300);
+        question.setTitle("my question title");
+        question.setVotes(3);
+        question.setQuestion("my question");
+
+        questionResource.createQuestion(new MockAuth(auth.getUserId()), question).toBlocking().singleOrDefault(null);
+
+        // then the question should be returned when asking for the users questions
+        List<Question> questions = questionResource.getQuestions(auth.getUserId()).toBlocking().single();
+        assertEquals(1, questions.size());
+        return questions.get(0);
+    }
+
+
+    private static Answer createAnswer() {
+        return createAnswer("this is an answer title", "this is the body of the answer");
+    }
+
+    private static Answer createAnswer(String title, String answerBody) {
+        Answer answer = new Answer();
+        answer.setAnswer(answerBody);
+        return answer;
     }
 
     private void addVote(Auth auth, Answer answer, int value) {
