+++ spec/src/test/java/impl/QuestionResourceTest.java
@@ -23,12 +23,19 @@ import se.fortnox.reactivewizard.jaxrs.WebException;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.function.BiFunction;
 
 import static impl.QuestionResourceImpl.FAILED_TO_SEARCH_FOR_QUESTIONS;
+import static impl.QuestionResourceImpl.INVALID_VOTE;
+import static impl.QuestionResourceImpl.QUESTION_NOT_FOUND;
+import static impl.TestSetup.getQuestion;
 import static impl.TestSetup.insertUser;
+import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
+import static io.netty.handler.codec.http.HttpResponseStatus.NOT_FOUND;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyString;
@@ -40,6 +47,7 @@ public class QuestionResourceTest {
     private static QuestionResource questionResource;
     private static AnswerResource   answerResource;
     private static UserResource     userResource;
+    private static QuestionVoteDao  questionVoteDao;
 
     @ClassRule
     public static PostgreSQLContainer postgreSQLContainer = new PostgreSQLContainer();
@@ -52,6 +60,7 @@ public class QuestionResourceTest {
         userResource = testSetup.getInjector().getInstance(UserResource.class);
         questionResource = testSetup.getInjector().getInstance(QuestionResource.class);
         answerResource = testSetup.getInjector().getInstance(AnswerResource.class);
+        questionVoteDao = testSetup.getInjector().getInstance(QuestionVoteDao.class);
     }
 
     @After
@@ -67,7 +76,7 @@ public class QuestionResourceTest {
     @Test
     public void shouldThrowErrorWhenServerIsDown() {
         QuestionDao          questionDao      = mock(QuestionDao.class);
-        QuestionResourceImpl questionResource = new QuestionResourceImpl(questionDao);
+        QuestionResourceImpl questionResource = new QuestionResourceImpl(questionDao, questionVoteDao);
         when(questionDao.getLatestQuestions(any())).thenReturn(error(new SQLException()));
 
         try {
@@ -81,7 +90,7 @@ public class QuestionResourceTest {
     @Test
     public void shouldBePossibleToGetQuestionBySlackThreadId() {
 
-        User createdUser = TestSetup.insertUser(userResource);
+        TestSetup.insertUser(userResource);
         Auth mockAuth    = newAuth();
 
         Question question      = TestSetup.getQuestion("my question title", "my question");
@@ -150,7 +159,7 @@ public class QuestionResourceTest {
 
     private Answer createAnswer(Auth mockAuth, long questionId, String answerBody) {
         Answer answer = TestSetup.getAnswer(answerBody);
-        answerResource.answerQuestion(mockAuth, answer, questionId).toBlocking().singleOrDefault(null);
+        answerResource.createAnswer(mockAuth, answer, questionId).toBlocking().singleOrDefault(null);
         return answer;
     }
 
@@ -206,11 +215,11 @@ public class QuestionResourceTest {
     @Test
     public void shouldBePossibleToSearchForAnswersThatAreNotCreatedByTheQuestionOwner() {
         // given two users
-        Auth firstUserAuth = createUserAndAuth();
+        Auth firstUserAuth  = createUserAndAuth();
         Auth secondUserAuth = createUserAndAuth();
 
         // and answers created both by the question owner and the other user
-        long questionId = createQuestion(firstUserAuth, "Question?","Question?").getId();
+        long questionId = createQuestion(firstUserAuth, "Question?", "Question?").getId();
         createAnswer(firstUserAuth, questionId, "1");
         Answer answerNotCreatedByOwner = createAnswer(secondUserAuth, questionId, "2");
 
@@ -226,7 +235,7 @@ public class QuestionResourceTest {
         // given that the query will fail
         QuestionDao questionDao = mock(QuestionDao.class);
         when(questionDao.getQuestions(anyString())).thenReturn(error(new WebException()));
-        QuestionResource questionResource = new QuestionResourceImpl(questionDao);
+        QuestionResource questionResource = new QuestionResourceImpl(questionDao, questionVoteDao);
 
         // when searching
         Observable<List<Question>> questions = questionResource.getQuestionsBySearchQuery("explode");
@@ -240,19 +249,195 @@ public class QuestionResourceTest {
             });
     }
 
-    private Auth createUserAndAuth() {
-        User createdUser = TestSetup.insertUser(userResource);
-        Auth mockAuth    = new MockAuth(createdUser.getId());
+    @Test
+    public void shouldBePossibleToAddQuestionAndFetchIt() {
+
+        User createdUser = insertUser(userResource);
+
+        // when question is created
+        Question question = getQuestion("my question title", "my question");
+        Auth     mockAuth = new MockAuth(createdUser.getId());
         mockAuth.setUserId(createdUser.getId());
-        return mockAuth;
+
+        // when we create the question
+        questionResource.createQuestion(mockAuth, question).toBlocking().singleOrDefault(null);
+
+        // then the question should be returned when asking for the users questions
+        List<Question> questions = questionResource.getQuestions(createdUser.getId()).toBlocking().single();
+        assertEquals(1, questions.size());
+
+        Question insertedQuestion = questions.get(0);
+        assertEquals("my question title", insertedQuestion.getTitle());
+        assertEquals("my question", insertedQuestion.getQuestion());
+        assertEquals(question.getBounty(), insertedQuestion.getBounty());
+        assertEquals(createdUser.getId(), insertedQuestion.getUserId());
+        assertNotNull(insertedQuestion.getId());
     }
 
-    private void generateQuestions(int questionsToGenerate) {
-        Auth mockAuth = createUserAndAuth();
+    @Test
+    public void shouldBePossibleToGetQuestionById() {
 
-        for (int i = 1; i <= questionsToGenerate; i++) {
-            createQuestion(mockAuth, "my question title " + i, "my question");
-        }
+        // when question is inserted
+        Question question = getQuestion("my question title", "my question");
+        Auth     mockAuth = newAuth();
+        mockAuth.setUserId(mockAuth.getUserId());
+        questionResource.createQuestion(mockAuth, question).toBlocking().singleOrDefault(null);
+        List<Question> questions = questionResource.getQuestions(mockAuth.getUserId()).toBlocking().single();
+        assertEquals(1, questions.size());
+
+        // then the question should be returned when asking for the specific question
+        Question selectedQuestion = questionResource.getQuestion(newAuth(), questions.get(0).getId()).toBlocking().single();
+
+        assertEquals("my question title", selectedQuestion.getTitle());
+        assertEquals("my question", selectedQuestion.getQuestion());
+        assertEquals(Integer.valueOf(300), selectedQuestion.getBounty());
+        assertEquals(mockAuth.getUserId(), selectedQuestion.getUserId().longValue());
+    }
+
+    @Test
+    public void shouldBePossibleToUpdateQuestion() {
+
+        User createdUser = insertUser(userResource);
+
+        // when question is inserted
+        Question question = getQuestion("my question title", "my question");
+
+        Auth mockAuth = new MockAuth(createdUser.getId());
+        mockAuth.setUserId(createdUser.getId());
+        questionResource.createQuestion(mockAuth, question).toBlocking().singleOrDefault(null);
+
+        Question storedQuestion = questionResource.getQuestions(createdUser.getId()).toBlocking().single().get(0);
+        storedQuestion.setBounty(400);
+        storedQuestion.setTitle("new title");
+        storedQuestion.setQuestion("new question body");
+        questionResource.updateQuestion(mockAuth, storedQuestion.getId(), storedQuestion).toBlocking().singleOrDefault(null);
+        List<Question> questions = questionResource.getQuestions(createdUser.getId()).toBlocking().single();
+        assertEquals(1, questions.size());
+
+        Question updatedQuestion = questions.get(0);
+        assertEquals("new title", updatedQuestion.getTitle());
+        assertEquals("new question body", updatedQuestion.getQuestion());
+        assertEquals(question.getBounty(), updatedQuestion.getBounty());
+        assertEquals(createdUser.getId(), updatedQuestion.getUserId());
+    }
+
+    @Test
+    public void shouldOnyReturnQuestionsForTheSpecifiedUser() {
+
+        User otherUser = insertUser(userResource);
+        User ourUser   = insertUser(userResource);
+
+        // when questions are inserted for different users
+        Question ourQuestion = getQuestion("our users question title", "our users question");
+
+        // when question is inserted
+        Question questionForOtherUser = getQuestion("other users question title", "other users question");
+
+        Auth auth = new MockAuth(ourUser.getId());
+        questionResource.createQuestion(auth, ourQuestion).toBlocking().singleOrDefault(null);
+        Auth authOtherUser = new MockAuth(otherUser.getId());
+        questionResource.createQuestion(authOtherUser, questionForOtherUser).toBlocking().singleOrDefault(null);
+
+        // then only questions for our user should be returned
+        List<Question> questions = questionResource.getQuestions(ourUser.getId()).toBlocking().single();
+        assertEquals(1, questions.size());
+
+        Question insertedQuestion = questions.get(0);
+        assertEquals("our users question title", insertedQuestion.getTitle());
+        assertEquals(ourUser.getId(), insertedQuestion.getUserId());
+    }
+
+    @Test
+    public void shouldBePossibleToDeleteQuestion() {
+
+        User createdUser = insertUser(userResource);
+
+        Auth mockAuth = new MockAuth(createdUser.getId());
+        mockAuth.setUserId(createdUser.getId());
+
+        // given questions exists
+        Question questionToInsert = getQuestion("my question title", "my question");
+
+        Question questionToInsert2 = getQuestion("my question title2", "my question2");
+
+        questionResource.createQuestion(mockAuth, questionToInsert).toBlocking().singleOrDefault(null);
+        questionResource.createQuestion(mockAuth, questionToInsert2).toBlocking().singleOrDefault(null);
+
+        // and the questions has answers
+        List<Question> questions = questionResource.getQuestions(createdUser.getId()).toBlocking().single();
+
+        Answer answer = new Answer();
+        answer.setAnswer("just a answer");
+        answerResource.createAnswer(mockAuth, answer, questions.get(0).getId()).toBlocking().singleOrDefault(null);
+        answerResource.createAnswer(mockAuth, answer, questions.get(1).getId()).toBlocking().singleOrDefault(null);
+
+        List<Question> questionsSaved = questionResource.getQuestions(createdUser.getId()).toBlocking().single();
+
+        // when we deletes a question
+        questionResource.deleteQuestion(mockAuth, questionsSaved.get(0).getId()).toBlocking().singleOrDefault(null);
+
+        // only the question we want to delete should be removed
+        List<Question> remaningQuestions = questionResource.getQuestions(createdUser.getId()).toBlocking().single();
+        assertThat(remaningQuestions.size()).isEqualTo(1);
+        assertThat(remaningQuestions.get(0).getTitle()).isEqualTo("my question title2");
+
+        // and the answers should be deleted as well for the deleted question
+        List<Answer> answersForTheNonDeletedQuestion = answerResource.getAnswers(mockAuth, questionsSaved.get(1).getId()).toBlocking().singleOrDefault(null);
+        assertThat(answersForTheNonDeletedQuestion).isNotEmpty();
+
+        List<Answer> answersForTheDeletedQuestion = answerResource.getAnswers(mockAuth, questionsSaved.get(0).getId()).toBlocking().singleOrDefault(null);
+        assertThat(answersForTheDeletedQuestion).isEmpty();
+    }
+
+    @Test
+    public void shouldNotDeleteQuestionThatDoesNotExist() {
+        User createdUser = insertUser(userResource);
+
+        Auth auth                  = new MockAuth(createdUser.getId());
+        long nonExistingQuestionId = 12;
+        assertThatExceptionOfType(WebException.class)
+            .isThrownBy(() -> questionResource.deleteQuestion(auth, nonExistingQuestionId).toBlocking().singleOrDefault(null))
+            .satisfies(e -> {
+                assertEquals(NOT_FOUND, e.getStatus());
+                assertEquals(QUESTION_NOT_FOUND, e.getError());
+            });
+    }
+
+    @Test
+    public void shouldYieldCorrectVotesWhenDownVotingAndUpVoting() {
+
+        Auth auth1 = newAuth();
+        Auth auth2 = newAuth();
+        Auth auth3 = newAuth();
+
+        Question question = questionResource.createQuestion(newAuth(), getQuestion("my question title", "my question")).toBlocking().singleOrDefault(null);
+
+        voteAndAssertSuccess(questionResource::downVoteQuestion, auth1, question, -1);
+        voteAndAssertFailure(questionResource::downVoteQuestion, auth1, question);
+        voteAndAssertSuccess(questionResource::upVoteQuestion, auth1, question, 0);
+        assertNoVote(auth1, question);
+        voteAndAssertSuccess(questionResource::upVoteQuestion, auth1, question, 1);
+        voteAndAssertFailure(questionResource::upVoteQuestion, auth1, question);
+
+        voteAndAssertSuccess(questionResource::upVoteQuestion, auth2, question, 2);
+
+        voteAndAssertSuccess(questionResource::upVoteQuestion, auth3, question, 3);
+
+        voteAndAssertSuccess(questionResource::downVoteQuestion, auth2, question, 2);
+        assertNoVote(auth2, question);
+
+        voteAndAssertSuccess(questionResource::downVoteQuestion, auth3, question, 1);
+        assertNoVote(auth2, question);
+    }
+
+    @Test
+    public void shouldThrowErrorWhenUserVotesOnOwnAnswer() {
+
+        Auth     auth     = newAuth();
+        Question question = questionResource.createQuestion(auth, getQuestion("my question title", "my question")).toBlocking().singleOrDefault(null);
+
+        voteAndAssertFailure(questionResource::downVoteQuestion, auth, question);
+        voteAndAssertFailure(questionResource::upVoteQuestion, auth, question);
     }
 
     @Test
@@ -272,9 +457,48 @@ public class QuestionResourceTest {
         }
     }
 
+    private void voteAndAssertSuccess(BiFunction<Auth, Long, Observable<Void>> call, Auth auth, Question question, int expectedVotes) {
+        call.apply(auth, question.getId()).test()
+            .awaitTerminalEvent()
+            .assertNoErrors();
+        assertThat(questionResource.getQuestionById(question.getId()).test().awaitTerminalEvent().getOnNextEvents())
+            .extracting(Question::getVotes)
+            .containsExactly(expectedVotes);
+    }
+
+    private void voteAndAssertFailure(BiFunction<Auth, Long, Observable<Void>> call, Auth auth, Question question) {
+        assertThatExceptionOfType(WebException.class)
+            .isThrownBy(() -> call.apply(auth, question.getId()).toBlocking().single())
+            .satisfies(e -> {
+                assertThat(e).isInstanceOf(WebException.class);
+                assertThat(e).hasFieldOrPropertyWithValue("status", BAD_REQUEST);
+                assertThat(e).hasFieldOrPropertyWithValue("error", INVALID_VOTE);
+            });
+    }
+
+    private void assertNoVote(Auth auth, Question answer) {
+        questionVoteDao.findVote(auth.getUserId(), answer.getId()).test().awaitTerminalEvent()
+            .assertNoValues();
+    }
 
     private Auth newAuth() {
         User createdUser = insertUser(userResource);
         return new MockAuth(createdUser.getId());
     }
+
+    private Auth createUserAndAuth() {
+        User createdUser = TestSetup.insertUser(userResource);
+        Auth mockAuth    = new MockAuth(createdUser.getId());
+        mockAuth.setUserId(createdUser.getId());
+        return mockAuth;
+    }
+
+    private void generateQuestions(int questionsToGenerate) {
+        Auth mockAuth = createUserAndAuth();
+
+        for (int i = 1; i <= questionsToGenerate; i++) {
+            createQuestion(mockAuth, "my question title " + i, "my question");
+        }
+    }
+
 }
